pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Denis Ivanov <dnb8866@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
pytest = "^8.3.3"
flake8 = "^7.1.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.8"
pytest-django = "^4.9.0"
django-extensions = "^3.2.3"
django-filter = "^24.3"
ipython = "^8.29.0"
django-bootstrap5 = "^24.3"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
